requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
Django==3.1.6
django-render-partial==0.4
django-smartfields==1.1.3
idna==2.10
isodate==0.6.0
lxml==4.6.2
Pillow==8.1.0
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.26.3
zeep==4.0.0
